Add TabbedPage for Map and Chat - Added Map page
